scripts/requirements.txt
alembic==1.12.1
aniso8601==9.0.1
annotated-types==0.6.0
apturl==0.5.2
arandr==0.1.10
argcomplete==1.8.1
asn1crypto==0.24.0
aspell-python-py3==1.15
async-timeout==4.0.3
attrs==23.1.0
axe-selenium-python==2.1.6
Babel==2.13.1
base58==2.1.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
bidict==0.22.1
binwalk==2.2.0
black==23.11.0
blinker==1.7.0
Brlapi==0.7.0
cachelib==0.9.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
cytoolz==0.11.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.14
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==2.4.2
docopt==0.6.2
dominate==2.9.0
duplicity==0.8.12.0
ecdsa==0.15
email-validator==2.1.0.post1
entrypoints==0.3
eth-abi==2.1.1
eth-hash==0.3.1
eth-typing==2.2.2
eth-utils==1.10.0
exceptiongroup==1.2.0
Faker==20.1.0
fasteners==0.14.1
feedparser==5.2.1
filelock==3.13.1
Flask==2.0.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Flash==2.0.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0
Flask-Logging==0.1.3
Flask-Login==0.6.3
flask-marshmallow==0.10.0
Flask-Migrate==4.0.5
Flask-Moment==1.0.5
flask-paginate==2023.10.24
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
flask-sqlalchemy==3.1.1
Flask-SSLify==0.1.5
flask-talisman==1.1.0
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
flask-wtf==1.2.1
future==0.18.2
gpg==1.13.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.12.0
httpcore==0.12.3
httplib2==0.14.0
httpx==0.16.1
hunspell==0.5.5
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflect==7.0.0
iniconfig==2.0.0
iotop==0.6
isodate==0.6.0
itsdangerous==2.1.2
Jinja2==2.11.3
keyring==18.0.1
keyrings.alt==3.4.0
kiwisolver==1.0.1
langid==1.1.6
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
limits==3.6.0
lockfile==0.12.2
louis==3.12.0
lxml==4.9.4
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
markdown-it-py==3.0.0
MarkupSafe==1.1.1
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
matplotlib==3.1.2
mdurl==0.1.2
monotonic==1.5
musicbrainzngs==0.7.1
mutagen==1.44.0
mypy-extensions==1.0.0
netifaces==0.10.4
numpy==1.17.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.46
openshot-qt==3.1.1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==23.2
paramiko==2.6.0
parsimonious==0.8.1
pathspec==0.11.2
pbr==6.0.0
pexpect==4.6.0
Pillow==7.0.0
pipreqs==0.4.13
platformdirs==4.0.0
pluggy==1.3.0
protobuf==3.6.1
psutil==5.5.1
py-cpuinfo==8.0.0
pycairo==1.16.2
pycountry==19.8.18
pycrypto==2.6.1
pycryptodome==3.10.1
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==2.5.2
pydantic-core==2.14.5
pydbus==0.6.0
pygments==2.17.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyparsing==2.4.6
pypresence==4.2.0
PyQt5==5.14.1
pyqtgraph==0.11.0rc0
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyserial==3.5
PySocks==1.6.8
pyspellchecker==0.7.2
pytest==7.4.3
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
python-engineio==4.8.0
python-magic==0.4.16
python-socketio==5.10.0
pytz==2023.3.post1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
quodlibet==4.2.1
redis==5.0.1
reportlab==3.5.34
requests==2.31.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
rich==13.7.0
scipy==1.3.3
SecretStorage==2.3.1
selenium==4.16.0
sentry-sdk==0.12.2
simple-websocket==1.0.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
speedtest-cli==2.1.2
SQLAlchemy==2.0.23
ssh-import-id==5.10
streamlink==1.5.0
style==1.1.0
system-service==0.3
systemd-python==234
testresources==2.0.1
tomli==2.0.1
toolz==0.11.1
trio==0.23.2
trio-websocket==0.11.1
tronpy==0.2.6
typing-extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
update==0.0.1
urllib3==2.1.0
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.24.7
visitor==0.1.3
wadllib==1.3.3
websocket-client==0.53.0
Werkzeug==2.0.1
wrapt==1.16.0
wsproto==1.2.0
wtforms==3.1.1
xkit==0.0.0
xxhash==2.0.2
yarg==0.1.9
youtube-dl==2020.3.24
zipp==3.17.0
zope.interface==4.7.1
